Adjusted console window behavior
Changed the way the console window is hidden to avoid using the registry.

Added console output for when the console window rmains open, this was added to the WPF close button "closing" action.  Output lets user know threads are closing and to wait a moment before the console will fully close.